Ignore RP-0 tooling costs in sim cost calculation
The upcoming RP-0 release adds a game mechanic called "tooling" that
aims to discourage the (ab)use of procedural parts for building
mission-specific rockets, and incentivize the reuse and incremental
evolution of major rocket components such as stages or launch vehicles.

Tooling accomplishes this by adding a large additional "tooling" cost to
each procedural part, to simulate the costs associated with setting up
the tooling (equipment, procedures, tools, etc.) needed to fabricate a
particular part. The player can pay this additional cost for each
procedural part, or pay a somewhat higher one-time fee to permanently
set up tooling for a specific procedural part with a specific shape and
size. Once the fee is paid, the player no longer needs to pay the
tooling cost for that part, significantly decreasing the cost. In
addition, the cost of parts with a similar shape and size, as well as
the cost to permanently set up tooling for similar parts, is greatly
reduced.

For example, prior to purchasing permanent tooling, a procedural level 1
tank with a diameter of 1 meter and length of 1 meter costs about 1000
funds, and the same tank with a diameter of 1 meter and length of 2
meters costs about 1011 funds. Purchasing tooling for the 1 meter tank
costs about 3994 funds, while tooling for the 2 meter tank costs around
4031 funds. After tooling for the 1 meter tank is purchased, its cost
drops to about 1 fund, while the cost of the 2 meter tank decreases to
about 144 funds. Additional tooling for the 2 meter tank can be
purchased for about 563 funds, after which the cost of the 2 meter tank
falls to about 3 funds. A 3 meter diameter, 1 meter long tank, however,
costs 4761 funds before purchasing toolings for the smaller tanks, and
4761 funds after.

While this mechanic works well for building rockets, the massive price
increase associated with untooled parts can result in a corresponding
massive price increase in the cost of simulations, since simulation
costs may be at least partially based on rocket costs. Because nothing
is actually being built when a simulation is run, tooling costs should
not count towards the rocket cost when simulation costs are being
calculated.

Compensating for tooling costs is fairly straightforward, in theory.
RP-0 uses a PartModule called ModuleTooling (and PartModules derived
from it) to calculate tooling costs and determine whether a given part
has permanent tooling purchased. The former is done through
GetModuleCost(), inherited from IPartCostModifier, and the latter
through the ModuleTooling-specific method IsUnlocked(). Compensating for
tooling costs could therefore be done in a relatively straightforward
manner using reflection:

    Type moduleToolingType;
    MethodInfo isUnlockedMI;
    foreach (var part in ship.parts)
    {
        foreach (var module in part)
        {
            if (moduleToolingType.IsInstanceOfType(module))
            {
                if (!isUnlockedMI.Invoke(module))
                {
                    // This is safe, since module must be ModuleTooling
                    // to have gotten here, and ModuleTooling derives
                    // from IPartCostModifier
                    toolingCost += ((IPartCostModifier)module).GetModuleCost(...);
                }
            }
        }
    }

Reflection is pretty slow, though, and can't be checked by the compiler.
Although this isn't being executed frequently, like the GUI code,
avoiding unnecessary performance hits is still preferable.

Delegates provide a nice alternative to reflection; they are usually
type-checked by the compiler, and their speed is on par with a regular
method call. Creating them just requires a function signature and a
MethodInfo for a method that matches the signature. Using plain
reflection would involve getting the right MethodInfo objects anyways,
so it isn't too much more work to convert them to delegates.

Ideally, the delegate would look like this:

    delegate bool IsUnlockedDelegate(ModuleTooling mt);

but because KRASH doesn't depend on RP-0, there's no way to declare a
ModuleTooling parameter. An untyped delegate could be used, but that
prevents compile-time type checking, so another solution is preferable.

Generics can be used to solve this problem, by using a type parameter T
as a stand-in for ModuleTooling. Constraining that type with the
classes/interfaces ModuleTooling derives from ensures that T resembles
ModuleTooling as much as possible to the compiler without actually
referencing it in code. The code then reduces to:

    delegate bool IsUnlockedDelegate(T t);
    IsUnlockedDelegate isUnlocked;
    foreach (var part in ship.parts)
    {
        foreach (var module in part)
        {
            if (module is T mt)
            {
                if (!isUnlocked(mt))
                {
                    toolingCost += module.GetModuleCost(...);
                }
            }
        }
    }

which is much closer to the code that would be written if ModuleTooling
could be referenced directly in code. Even if it's not significantly
shorter, hopefully it's closer to idiomatic C# code, and is easier to
maintain.

The final step is to create a ModuleTooling-specific version of the code
using Type.MakeGenericType at runtime, then grab the cost-calculating
code and turn it into a delegate.

The rest of the code is setup and error handling code, taking care of
finding the right methods, converting them into delegates, and creating
the ModuleTooling-specific version of RP0Helper.

The decision to use a null check for the RP-0 assembly, but use a
try-catch for everything else, is based on the assumption that it's
quite likely for the RP-0 DLL to be missing, so a regular if statement
is preferable to throwing/catching an exception. If the RP-0 DLL is
found, the rest of the code is expected to succeed, since the API
is assumed to not change frequently, so a try-catch is more appropriate,
and less noisy.